moved rebin tests from dataset to variable
