+++ variable/rebin.cpp
@@ -31,6 +31,7 @@ void rebin_non_inner(const Dim dim, const VariableConstView &oldT,
   int iold = 0;
   int inew = 0;
   const bool is_bool = newT.dtype() == dtype<bool>;
+  // const bool is_bool = newT.dtype().index == std::type_index(typeid(bool));
   while ((iold < oldSize) && (inew < newSize)) {
     auto xo_low = xold[iold];
     auto xo_high = xold[iold + 1];
